+++ R/tri_mesh.r
@@ -3,7 +3,12 @@ path2seg <- function(x) {
   head(suppressWarnings(matrix(x, nrow = length(x) + 1, ncol = 2, byrow = FALSE)), -2L)
 }
 
-
+#' Deprecated from rangl
+#' @rdname rangl-deprecated
+#' @export
+tri_mesh <- function(x, ...) {
+  .Deprecated("mesh", package= "rangl", old = "tri_mesh")
+}
 
 tri_mesh_map_table1 <- function(tabs, max_area = NULL) {
   tabs$v$countingIndex <- seq(nrow(tabs$v))